feat: dashbaord server para montar logs e metricas;feat: frontend minimo estatico para exibir um painel de observabilidade

+++ docker/database/init.sql
@@ -8,7 +8,7 @@ CREATE TABLE event_logs (
     amount DECIMAL(10,2) NOT NULL,
     event_timestamp TIMESTAMP NOT NULL,
     instance_id VARCHAR(50) NOT NULL,
-    latency_ms INTEGER NOT NULL,
+    latency_ms BIGINT NOT NULL,
     received_at TIMESTAMP NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
